% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModGlossary.R
\name{srvModGlossary}
\alias{srvModGlossary}
\title{Server Module for the Glossary}
\usage{
srvModGlossary(input, output, session)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}
}
\description{
Server Module for the Glossary
}
