% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModGraph.R
\name{srvModGraph}
\alias{srvModGraph}
\title{Server Module for Result Graph}
\usage{
srvModGraph(input, output, session, orders = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{orders}{to plot the graph}
}
\description{
Server Module for Result Graph
}
