% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModInitialValues.R
\name{srvModInitialValues}
\alias{srvModInitialValues}
\title{Server Module for Initial Values}
\usage{
srvModInitialValues(input, output, session, initial = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{initial}{initial values}
}
\value{
initial_rv
}
\description{
Server Module for Initial Values
}
