% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModInitialValuesCSV.R
\name{srvModInitialValuesCSV}
\alias{srvModInitialValuesCSV}
\title{Server Module for Initial Values CSV file Down/Up-load}
\usage{
srvModInitialValuesCSV(input, output, session, initial = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{initial}{initial values}
}
\value{
initial_rv
}
\description{
Server Module for Initial Values CSV file Down/Up-load
}
