% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModInitialValuesExcel.R
\name{srvModInitialValuesExcel}
\alias{srvModInitialValuesExcel}
\title{Server Module for Initial Values Excel Down/Up-load}
\usage{
srvModInitialValuesExcel(input, output, session, initial = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{initial}{initial values}
}
\value{
initial_rv
}
\description{
Server Module for Initial Values Excel Down/Up-load
}
