% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModOrders.R
\name{srvModOrders}
\alias{srvModOrders}
\title{Server Module for Orders}
\usage{
srvModOrders(input, output, session, orders = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{orders}{to be shown in a table}
}
\description{
Server Module for Orders
}
