% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModParam.R
\name{srvModParam}
\alias{srvModParam}
\title{Server Module for Parameters}
\usage{
srvModParam(input, output, session)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}
}
\value{
param_rv
}
\description{
Server Module for Parameters
}
