% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModPlay.R
\name{srvModPlay}
\alias{srvModPlay}
\title{Server Module Play}
\usage{
srvModPlay(input, output, session, initial = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}
}
\value{
customer_rv
}
\description{
Server Module Play
}
