% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModResult.R
\name{srvModResult}
\alias{srvModResult}
\title{Server Module for Results}
\usage{
srvModResult(input, output, session, result = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{values}{}
}
\description{
Server Module for Results
}
