% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiModAbout.R
\name{uiModAbout}
\alias{uiModAbout}
\title{User Interface Module for the About}
\usage{
uiModAbout(id, title = id, ..., value = title)
}
\arguments{
\item{id}{character, used to identify a namespace
other parameters , see \code{shiny::\link[shiny]{tabPanel}}}
}
\value{
a \code{shiny::\link[shiny]{tabPanel}} containing UI elements
}
\description{
User Interface Module for the About
}
