% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiModInitialValuesCSV.R
\name{uiModInitialValuesCSV}
\alias{uiModInitialValuesCSV}
\title{User Interface Module for Initial Values CSV file Down/Up-load}
\usage{
uiModInitialValuesCSV(id, title = id, ..., value = title)
}
\arguments{
\item{id}{character, used to identify a namespace
other parameters , see \code{shiny::\link[shiny]{tabPanel}}}
}
\value{
a \code{shiny::\link[shiny]{tabPanel}} containing UI elements
}
\description{
User Interface Module for Initial Values CSV file Down/Up-load
}
