% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiModInitialValuesExcel.R
\name{uiModInitialValuesExcel}
\alias{uiModInitialValuesExcel}
\title{User Interface Module for Initial Values Excel Down/Up-load}
\usage{
uiModInitialValuesExcel(id, title = id, ..., value = title)
}
\arguments{
\item{id}{character, used to identify a namespace
other parameters , see \code{shiny::\link[shiny]{tabPanel}}}
}
\value{
a \code{shiny::\link[shiny]{tabPanel}} containing UI elements
}
\description{
User Interface Module for Initial Values Excel Down/Up-load
}
