% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvModCompany.R
\name{srvModCompany}
\alias{srvModCompany}
\title{Server Module for a Company}
\usage{
srvModCompany(input, output, session, inputs = NULL, demand = NULL,
  supply = NULL)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{demand}{demand company}

\item{supply}{supply company}

\item{initial}{initial values}
}
\value{
company_rv
}
\description{
Server Module for a Company
}
